remote_cs04/for_aws/requirements.txt

Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-SQLAlchemy==2.4.1
dataset==1.1.2
rope==0.14.0
